pyproject.toml
[project]
name = "case-studies-final-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imblearn>=0.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "psycopg>=3.3.3",
    "psycopg2>=2.9.11",
    "pyhealth",
    "pyodbc>=5.3.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.3",
]
